Refactor application structure and implement new widget system
- Removed the `open.rs` file and its associated functionality.
- Consolidated library path validation logic into `validate_library_path` function in `validate/app.rs`.
- Introduced a new default configuration for the application in `default/app.rs`.
- Added a new `pipeline` module to manage application pipelines.
- Created a new `widget` module to handle UI components, including a filter widget.
- Updated the `main.rs` file to integrate the new widget system.
- Refactored the `Filter` struct to manage book filtering and highlights.
- Implemented a new `Open` widget to handle book opening functionality.
- Enhanced error handling and logging throughout the application.
- Updated various data structures and types for improved clarity and consistency.

+++ src/config/ui/filter/table.rs
@@ -16,7 +16,7 @@ pub struct Column {
     pub label_fg: String,
 
     pub ratio: u16,
-    pub position: i32,
+    pub position: u16,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
